rename mulighetsrommet-api -> client
