--- frontend/mulighetsrommet-api/package.json
@@ -1,13 +1,0 @@
-{
-  "private": true,
-  "name": "mulighetsrommet-api",
-  "version": "0.1.0",
-  "module": "src/index.ts",
-  "types": "src/index.ts",
-  "scripts": {
-    "generate": "openapi --useOptions --input ../../mulighetsrommet-api/src/main/resources/web/openapi.yml --output ./src"
-  },
-  "devDependencies": {
-    "openapi-typescript-codegen": "^0.20.1"
-  }
-}
