--- frontend/mulighetsrommet-api/src/index.ts
@@ -1,14 +1,0 @@
-/* istanbul ignore file */
-/* tslint:disable */
-/* eslint-disable */
-export { ApiError } from './core/ApiError';
-export { CancelablePromise } from './core/CancelablePromise';
-export { OpenAPI } from './core/OpenAPI';
-
-export type { Innsatsgruppe } from './models/Innsatsgruppe';
-export type { Tiltaksgjennomforing } from './models/Tiltaksgjennomforing';
-export { Tiltakskode } from './models/Tiltakskode';
-export type { Tiltakstype } from './models/Tiltakstype';
-
-export { InternalService } from './services/InternalService';
-export { MulighetsrommetService } from './services/MulighetsrommetService';
