--- frontend/mulighetsrommet-api/src/models/Innsatsgruppe.ts
@@ -1,9 +1,0 @@
-/* istanbul ignore file */
-/* tslint:disable */
-/* eslint-disable */
-
-export type Innsatsgruppe = {
-    id: number;
-    tittel: string;
-    beskrivelse: string;
-}
