--- frontend/mulighetsrommet-api/src/models/Tiltakskode.ts
@@ -1,110 +1,0 @@
-/* istanbul ignore file */
-/* tslint:disable */
-/* eslint-disable */
-
-export enum Tiltakskode {
-    ABIST = 'ABIST',
-    ABOPPF = 'ABOPPF',
-    ABTBOPPF = 'ABTBOPPF',
-    ABUOPPF = 'ABUOPPF',
-    AMBF1 = 'AMBF1',
-    AMBF2 = 'AMBF2',
-    AMBF3 = 'AMBF3',
-    AMO = 'AMO',
-    AMOB = 'AMOB',
-    AMOE = 'AMOE',
-    AMOY = 'AMOY',
-    ANNUTDANN = 'ANNUTDANN',
-    ARBDOGNSM = 'ARBDOGNSM',
-    ARBFORB = 'ARBFORB',
-    ARBRDAGSM = 'ARBRDAGSM',
-    ARBRRDOGN = 'ARBRRDOGN',
-    ARBRRHBAG = 'ARBRRHBAG',
-    ARBRRHBSM = 'ARBRRHBSM',
-    ARBRRHDAG = 'ARBRRHDAG',
-    ARBTREN = 'ARBTREN',
-    ASV = 'ASV',
-    ATG = 'ATG',
-    AVKLARAG = 'AVKLARAG',
-    AVKLARKV = 'AVKLARKV',
-    AVKLARSP = 'AVKLARSP',
-    AVKLARSV = 'AVKLARSV',
-    AVKLARUS = 'AVKLARUS',
-    BIA = 'BIA',
-    BIO = 'BIO',
-    BREVKURS = 'BREVKURS',
-    DIGIOPPARB = 'DIGIOPPARB',
-    DIVTILT = 'DIVTILT',
-    EKSPEBIST = 'EKSPEBIST',
-    ENKELAMO = 'ENKELAMO',
-    ENKFAGYRKE = 'ENKFAGYRKE',
-    ETAB = 'ETAB',
-    FLEKSJOBB = 'FLEKSJOBB',
-    FORSAMOENK = 'FORSAMOENK',
-    FORSAMOGRU = 'FORSAMOGRU',
-    FORSFAGENK = 'FORSFAGENK',
-    FORSFAGGRU = 'FORSFAGGRU',
-    FORSHOYUTD = 'FORSHOYUTD',
-    FUNKSJASS = 'FUNKSJASS',
-    GRUFAGYRKE = 'GRUFAGYRKE',
-    GRUNNSKOLE = 'GRUNNSKOLE',
-    GRUPPEAMO = 'GRUPPEAMO',
-    HOYEREUTD = 'HOYEREUTD',
-    HOYSKOLE = 'HOYSKOLE',
-    INDJOBSTOT = 'INDJOBSTOT',
-    INDOPPFAG = 'INDOPPFAG',
-    INDOPPFOLG = 'INDOPPFOLG',
-    INDOPPFSP = 'INDOPPFSP',
-    INDOPPRF = 'INDOPPRF',
-    INKLUTILS = 'INKLUTILS',
-    INST_S = 'INST_S',
-    IPSUNG = 'IPSUNG',
-    ITGRTILS = 'ITGRTILS',
-    JOBBBONUS = 'JOBBBONUS',
-    JOBBFOKUS = 'JOBBFOKUS',
-    JOBBK = 'JOBBK',
-    JOBBKLUBB = 'JOBBKLUBB',
-    JOBBSKAP = 'JOBBSKAP',
-    KAT = 'KAT',
-    KURS = 'KURS',
-    LONNTIL = 'LONNTIL',
-    LONNTILAAP = 'LONNTILAAP',
-    LONNTILL = 'LONNTILL',
-    LONNTILS = 'LONNTILS',
-    MENTOR = 'MENTOR',
-    MIDLONTIL = 'MIDLONTIL',
-    NETTAMO = 'NETTAMO',
-    NETTKURS = 'NETTKURS',
-    OPPLT2AAR = 'OPPLT2AAR',
-    PRAKSKJERM = 'PRAKSKJERM',
-    PRAKSORD = 'PRAKSORD',
-    PV = 'PV',
-    REAKTUFOR = 'REAKTUFOR',
-    REFINO = 'REFINO',
-    SPA = 'SPA',
-    STATLAERL = 'STATLAERL',
-    SUPPEMP = 'SUPPEMP',
-    SYSSLANG = 'SYSSLANG',
-    SYSSOFF = 'SYSSOFF',
-    TIDSUBLONN = 'TIDSUBLONN',
-    TILPERBED = 'TILPERBED',
-    TILRETTEL = 'TILRETTEL',
-    TILRTILSK = 'TILRTILSK',
-    TILSJOBB = 'TILSJOBB',
-    UF_REPENL_ = 'UFØREPENLØ',
-    UTBHLETTPS = 'UTBHLETTPS',
-    UTBHPSLD = 'UTBHPSLD',
-    UTBHSAMLI = 'UTBHSAMLI',
-    UTDPERMVIK = 'UTDPERMVIK',
-    UTDYRK = 'UTDYRK',
-    UTVAOONAV = 'UTVAOONAV',
-    UTVOPPFOPL = 'UTVOPPFOPL',
-    VALS = 'VALS',
-    VARLONTIL = 'VARLONTIL',
-    VASV = 'VASV',
-    VATIAROR = 'VATIAROR',
-    VIDRSKOLE = 'VIDRSKOLE',
-    VIKARBLED = 'VIKARBLED',
-    VV = 'VV',
-    YHEMMOFF = 'YHEMMOFF',
-}
