+++ frontend/mulighetsrommet-veileder-flate/package.json
@@ -48,7 +48,7 @@
     "jotai": "^1.5.3",
     "lodash.throttle": "^4.1.1",
     "mem": "^9.0.2",
-    "mulighetsrommet-api": "*",
+    "mulighetsrommet-api-client": "*",
     "node-forge": "^1.3.0",
     "nth-check": "^2.0.1",
     "prop-types": "^15.7.2",
