+++ frontend/mulighetsrommet-veileder-flate/src/hooks/tiltaksgjennomforing/useTiltaksgjennomforing.ts
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query';
-import { MulighetsrommetService, Tiltaksgjennomforing } from 'mulighetsrommet-api';
+import { MulighetsrommetService, Tiltaksgjennomforing } from 'mulighetsrommet-api-client';
 import { QueryKeys } from '../../core/api/QueryKeys';
 
 export default function useTiltaksgjennomforing(id: number) {
