+++ frontend/mulighetsrommet-veileder-flate/src/hooks/tiltaksgjennomforing/useTiltaksgjennomforingerByTiltakskode.ts
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query';
-import { MulighetsrommetService, Tiltaksgjennomforing, Tiltakskode } from 'mulighetsrommet-api';
+import { MulighetsrommetService, Tiltaksgjennomforing, Tiltakskode } from 'mulighetsrommet-api-client';
 import { QueryKeys } from '../../core/api/QueryKeys';
 
 export default function useTiltaksgjennomforingerByTiltakskode(tiltakskode: Tiltakskode) {
