+++ frontend/mulighetsrommet-veileder-flate/src/hooks/tiltakstype/useTiltakstype.ts
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query';
-import { MulighetsrommetService, Tiltakskode, Tiltakstype } from 'mulighetsrommet-api';
+import { MulighetsrommetService, Tiltakskode, Tiltakstype } from 'mulighetsrommet-api-client';
 import { QueryKeys } from '../../core/api/QueryKeys';
 
 export default function useTiltakstype(tiltakskode: Tiltakskode) {
