+++ frontend/mulighetsrommet-veileder-flate/src/mock/entities/tiltakstype.ts
@@ -1,5 +1,5 @@
 import { Entity } from '@mswjs/data/lib/glossary';
-import { Tiltakskode, Tiltakstype } from 'mulighetsrommet-api';
+import { Tiltakskode, Tiltakstype } from 'mulighetsrommet-api-client';
 import { DatabaseDictionary } from '../database';
 
 export type TiltakstypeEntity = Entity<DatabaseDictionary, 'tiltakstype'>;
