+++ frontend/mulighetsrommet-veileder-flate/src/mock/fixtures/tiltakstyper.ts
@@ -1,4 +1,4 @@
-import { Tiltakskode, Tiltakstype } from 'mulighetsrommet-api';
+import { Tiltakskode, Tiltakstype } from 'mulighetsrommet-api-client';
 export const tiltakstyper: Tiltakstype[] = [
   {
     id: 1,
